backend/requirements.txt
attrs==25.4.0
bcrypt==4.1.3
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
outcome==1.3.0.post0
passlib==1.7.4
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.1.1
PySocks==1.7.1
python-dotenv==1.0.1
selenium==4.38.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.5.0
websocket-client==1.9.0
Werkzeug==3.1.4
wsproto==1.3.2
